Django DRF and DevExtreme DataGrid test

+++ attendees/whereabouts/migrations/0008_room.py
@@ -24,7 +24,7 @@ class Migration(migrations.Migration):
                 ('display_name', models.CharField(db_index=True, max_length=50)),
                 ('slug', models.SlugField(max_length=50, unique=True)),
                 ('label', models.CharField(blank=True, max_length=20)),
-                ('accessibility', models.IntegerField(default=0)),
+                ('accessibility', models.SmallIntegerField(default=0)),
                 ('suite', models.ForeignKey(null=True, on_delete=django.db.models.deletion.SET_NULL, to='whereabouts.Suite')),
             ],
             options={
